refactor(#48): refatora tipagens das páginas de session
