+++ src/pages/Sessions/index.d.ts
@@ -1,0 +1,20 @@
+export type SessionType = "Treino" | "Jogo";
+
+export type SessionData = Session[];
+
+export type Session = {
+  id: string;
+  type: SessionType;
+  date: string;
+  local: string;
+  description?: string;
+  opponent?: string;
+}
+
+export type SessionMeta = {
+  type: SessionType;
+  date: string;
+  local: string;
+  description?: string;
+  opponent?: string;
+};
